[FEATURE] Added the reading of the cc report files (#3)
## 1) Description

Added the necessary files to read the cc report files produced by radon.
Added a development dependency to generate code to that can ensure the
data structure is correct.

## 2) Technical choice
TS cast are unsafe, and I didn't want to manually check if the types
were correct. This is why I added the dependency.
The tests are fairly simple as there is much to test except that it in
fact does read the file and find that the types are good.