[BUG] Adjusted workflows (#51)
## 1) Description

Updated the triggers of the workflows to not run on pushes.
Updated the Dependabot schedule from daily to weekly. 

This change is quite important as a bad run prevents #50 to be merged.

## 2) Technical choice

- The action is only built to be run on PR there is no sense to test it
on pushes. At least for now.
- Running the code QL analysis on each push is a waste of resources,
once on PR and on cron schedule is enough as all meaningful pushes were
PR and thus analyzed.
- Updating dependencies on a weekly is enough. I mean, we don't even do
releases this often.

## 3) Checks

- [X] My code follows the contributing guidelines
- [X] I have read the code of conduct
- [X] I have performed a self-review of my code
- [X] My changes generate no new warnings
- [X] New and existing unit tests pass locally with my changes